pyproject.toml
[tool.poetry]
name = "blog-flask-pydantic-pytest"
version = "0.1.0"
description = "A Flask blog app that uses pydantic for validation and pytest for testing."
authors = ["Danila Kritsky <danilakritsky@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
pydantic = {extras = ["email"], version = "^1.9.2"}

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
mypy = "^0.971"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
ipython = "^8.4.0"
flake8 = "^5.0.4"
python-dotenv = "^0.20.0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov --cov-report term-missing"

[tool.black]
preview = true
line-length = 79

[tool.mypy]
show_error_codes = true